FEAT(Gruntfile): set up grunt for more robust testing using mocha
yeoman test helpers are useful for testing the actual generation of
code but not so useful for testing whether the generated app works
as expected. for that, we build the app in a temp/ dir and run the
app's own suite of tests to make sure it works as expected.  the app
has not yet been moved into this generator but grunt and the test
suite are now properly set up to test it.