FEAT(env/tokens): add env vars and tokens and API keys
* Add prompts and logic for external service prompts such
including Amazon API keys and a MONGOLAB_URL.  @wip: Additional
options should be added once the app is fully functional.

* add in secretGenerator for generating an APP_SECRET and a
TOKEN_SECRET to be used by express so this does not have to
be done each time.

* change _processDirectory method in app/index.js to make a call
to this.template() without including a third context parameter.
This way, yeoman will use this (the generator itself) as the
context, giving it access to all of the necessary config settings.