FEAT(deploy/index.js): add ability to specify organization/repo
yo leviathan:deploy now allows the user to specify an organization
to use when creating the repo on github as well as a custom repo
name.  Simply use the format organization_name/repo_name when
prompted for a github repo name by the generator.  default remains
the name of the cwd.